refactor: break out admin js
